Add dedicated type file for future translations
