+++ resources/js/pages/auth/Login.vue
@@ -9,7 +9,7 @@ import InputText from 'primevue/inputtext';
 import Button from 'primevue/button';
 import Password from 'primevue/password';
 import Message from 'primevue/message';
-import { LoginTranslations } from '@/types';
+import { LoginTranslations } from '@/types/translations';
