+++ resources/js/types/index.d.ts
@@ -35,17 +35,4 @@ export interface User {
     updated_at: string;
 }
 
-export interface LoginTranslations {
-    title:        string;
-    description:  string;
-    simple_title: string;
-    email:        string;
-    password:     string;
-    remember_me:  string;
-    forgot_pass:  string;
-    no_account:   string;
-    sign_up:      string;
-}
-
-
 export type BreadcrumbItemType = BreadcrumbItem;
