+++ resources/js/types/translations.d.ts
@@ -1,0 +1,11 @@
+export interface LoginTranslations {
+    title:        string;
+    description:  string;
+    simple_title: string;
+    email:        string;
+    password:     string;
+    remember_me:  string;
+    forgot_pass:  string;
+    no_account:   string;
+    sign_up:      string;
+}
